requirements.txt
openai-whisper==20230314

pyannote.audio~=2.1.1
pyannote.core~=4.5
pyannote.database~=4.1.3
pyannote.metrics~=3.2.1
pyannote.pipeline~=2.3

setuptools~=65.6.3
setuptools-rust~=1.5.2

tqdm>=4.65.0

gradio~=3.36.1
gradio-client~=0.2.7

# add pytorch to override the one installed by pyannote.audio

torch~=1.11.0
torchvision~=0.12.0
torchaudio~=0.11.0
#optional: 
#sphinx~=5.0.2 


